added week-4 day-5 codes

+++ Week-4/Day-5/C_Number_of_Segments_with_Small_Sum.cpp
@@ -18,7 +18,7 @@ int main()
     for(int i=0;i<n;i++){
         cin>>v[i];
     }
-    int l=0,r=0,ans=0;
+    ll l=0,r=0,ans=0;
     ll sum=0;
     while(n>r){
         sum+=v[r];
